Trait moved to class
